feat(draft): zod 추가 및 API 유효성 스키마 변환 유틸 추가
